+++ starsim/__init__.py
@@ -2,7 +2,7 @@ from .version import __version__, __versiondate__, __license__
 from .settings      import *
 from .parameters    import *
 from .utils         import *
-from .dists         import *
+from .distributions import *
 from .states        import *
 from .people        import *
 from .modules       import *
@@ -19,10 +19,14 @@ from .sim           import *
 from .run           import *
 from .samples       import *
 
+# Assign the root folder
 import sciris as sc
 root = sc.thispath(__file__).parent
-del sc # Don't keep this in the module
 
 # Import the version and print the license
 if options.verbose:
     print(__license__)
+
+# Double-check key requirements -- should match setup.py
+sc.require(['sciris>=3.1.5', 'pandas>=2.0.0', 'scipy', 'numba', 'networkx'], message=f'The following dependencies for Starsim {__version__} were not met: <MISSING>.')
+del sc # Don't keep this in the module