+++ starsim/disease.py
@@ -282,14 +282,21 @@ class Infection(Disease):
                     continue
 
                 # Calculate probability of a->b transmission.
-                beta_per_dt = net.beta_per_dt(disease_beta=beta, dt=people.dt)
+                beta_per_dt = net.beta_per_dt(disease_beta=beta, dt=people.dt) # TODO: should this be sim.dt?
                 p_transmit = rel_trans[src] * rel_sus[trg] * beta_per_dt
 
                 slots_s = people.slot[src] # Slots for the possible source
                 slots_t = people.slot[trg] # Slots for the possible target
-                rvs_s = self.rng_source.rvs(size=np.max(slots_s) + 1)[slots_s]
-                rvs_t = self.rng_target.rvs(size=np.max(slots_t) + 1)[slots_t]
-                rvs = np.remainder(rvs_s + rvs_t, 1) # Generate a new random number based on the two other random numbers
+                # slots = (slots_s + slots_t) % sim.pars.n_agents # CK: faster, but some pairs would get the same draws
+                # rvs = self.rng_source.rvs(size=np.max(slots)+1)[slots]
+                rvs_s = self.rng_source.rvs(n=np.max(slots_s) + 1)[slots_s]
+                rvs_t = self.rng_target.rvs(n=np.max(slots_t) + 1)[slots_t]
+                
+                # Generate a new random number based on the two other random numbers -- 3x faster than `rvs = np.remainder(rvs_s + rvs_t, 1)`
+                rvs = rvs_s + rvs_t
+                inds = np.where(rvs>1.0)[0]
+                rvs[inds] -= 1
+                
                 new_cases_bool = rvs < p_transmit
                 new_cases.append(trg[new_cases_bool])
                 sources.append(src[new_cases_bool])