+++ starsim/diseases/ncd.py
@@ -5,7 +5,6 @@ Define non-communicable disease (NCD) model
 import numpy as np
 import starsim as ss
 import sciris as sc
-import scipy.stats as sps
 
 __all__ = ['NCD']
 
@@ -22,7 +21,7 @@ class NCD(ss.Disease):
             initial_risk = ss.bernoulli(p=0.3), # Initial prevalence of risk factors
             #'affection_rate': ss.rate(p=0.1), # Instantaneous rate of acquisition applied to those at risk (units are acquisitions / year)
             dur_risk = ss.expon(scale=10),
-            prognosis = sps.weibull_min(c=2, scale=5), # Time in years between first becoming affected and death
+            prognosis = ss.weibull(c=2, scale=5), # Time in years between first becoming affected and death
         )
 
         super().__init__(ss.omerge(default_pars, pars))
@@ -48,7 +47,7 @@ class NCD(ss.Disease):
         alive_uids = ss.true(sim.people.alive)
         initial_risk = self.pars['initial_risk'].filter(alive_uids)
         self.at_risk[initial_risk] = True
-        self.ti_affected[initial_risk] = sim.ti + sc.randround(self.pars['dur_risk'].urvs(initial_risk) / sim.dt)
+        self.ti_affected[initial_risk] = sim.ti + sc.randround(self.pars['dur_risk'].rvs(initial_risk) / sim.dt)
         return initial_risk
 
     def update_pre(self, sim):
@@ -61,7 +60,7 @@ class NCD(ss.Disease):
     def make_new_cases(self, sim):
         new_cases = ss.true(self.ti_affected == sim.ti)
         self.affected[new_cases] = True
-        prog_years = self.pars.prognosis.urvs(new_cases)
+        prog_years = self.pars.prognosis.rvs(new_cases)
         self.ti_dead[new_cases] = sim.ti + sc.randround(prog_years / sim.dt)
         super().set_prognoses(sim, new_cases)
         return new_cases