+++ starsim/diseases/syphilis.py
@@ -45,11 +45,11 @@ class Syphilis(ss.Infection):
         # Parameters
         default_pars = dict(
             # Adult syphilis natural history, all specified in years
-            dur_exposed = ss.lognorm_o(mean=1 / 12, stdev=1 / 36),  # https://pubmed.ncbi.nlm.nih.gov/9101629/
-            dur_primary = ss.lognorm_o(mean=1.5 / 12, stdev=1 / 36),  # https://pubmed.ncbi.nlm.nih.gov/9101629/
+            dur_exposed = ss.lognorm_ex(mean=1 / 12, stdev=1 / 36),  # https://pubmed.ncbi.nlm.nih.gov/9101629/
+            dur_primary = ss.lognorm_ex(mean=1.5 / 12, stdev=1 / 36),  # https://pubmed.ncbi.nlm.nih.gov/9101629/
             dur_secondary = ss.normal(loc=3.6 / 12, scale=1.5 / 12),  # https://pubmed.ncbi.nlm.nih.gov/9101629/
-            dur_latent_temp = ss.lognorm_o(mean=1, stdev=6 / 12),  # https://pubmed.ncbi.nlm.nih.gov/9101629/
-            dur_latent_long = ss.lognorm_o(mean=20, stdev=8),  # https://pubmed.ncbi.nlm.nih.gov/9101629/
+            dur_latent_temp = ss.lognorm_ex(mean=1, stdev=6 / 12),  # https://pubmed.ncbi.nlm.nih.gov/9101629/
+            dur_latent_long = ss.lognorm_ex(mean=20, stdev=8),  # https://pubmed.ncbi.nlm.nih.gov/9101629/
             p_latent_temp = ss.bernoulli(p=0.25),  # https://pubmed.ncbi.nlm.nih.gov/9101629/
             p_tertiary = ss.bernoulli(p=0.35),  # https://www.ncbi.nlm.nih.gov/pmc/articles/PMC4917057/
 
@@ -200,11 +200,11 @@ class Syphilis(ss.Infection):
 
         # Set future dates and probabilities
         # Exposed to primary
-        dur_exposed = self.pars.dur_exposed.urvs(uids)
+        dur_exposed = self.pars.dur_exposed.rvs(uids)
         self.ti_primary[uids] = sim.ti + rr(dur_exposed / sim.dt)
 
         # Primary to secondary
-        dur_primary = self.pars.dur_primary.urvs(uids)
+        dur_primary = self.pars.dur_primary.rvs(uids)
         self.ti_secondary[uids] = self.ti_primary[uids] + rr(dur_primary / sim.dt)
 
         return
@@ -212,10 +212,10 @@ class Syphilis(ss.Infection):
     def set_secondary_prognoses(self, sim, uids):
         """ Set prognoses for people who have just progressed to secondary infection """
 
-        dur_secondary = self.pars.dur_secondary.urvs(uids)
+        dur_secondary = self.pars.dur_secondary.rvs(uids)
 
         # Secondary to latent_temp or latent_long
-        latent_temp = self.pars.p_latent_temp.urvs(uids)
+        latent_temp = self.pars.p_latent_temp.rvs(uids)
         latent_temp_uids = uids[latent_temp]
         latent_long_uids = uids[~latent_temp]
 
@@ -229,20 +229,20 @@ class Syphilis(ss.Infection):
 
     def set_latent_temp_prognoses(self, sim, uids):
         # Primary to secondary
-        dur_latent_temp = self.pars.dur_latent_temp.urvs(uids)
+        dur_latent_temp = self.pars.dur_latent_temp.rvs(uids)
         self.ti_secondary[uids] = self.ti_latent_temp[uids] + rr(dur_latent_temp / sim.dt)
         return
 
     def set_latent_long_prognoses(self, sim, uids):
 
-        dur_latent = self.pars.dur_latent_long.urvs(uids)
+        dur_latent = self.pars.dur_latent_long.rvs(uids)
 
         # Primary to secondary
         dur_latent_long = dur_latent
         self.ti_secondary[uids] = self.ti_latent_temp[uids] + rr(dur_latent_long / sim.dt)
 
         # Latent_long to tertiary
-        tertiary = self.pars.p_tertiary.urvs(uids)
+        tertiary = self.pars.p_tertiary.rvs(uids)
         tertiary_uids = uids[tertiary]
         self.ti_tertiary[tertiary_uids] = self.ti_latent_long[tertiary_uids] + rr(dur_latent_long[tertiary] / sim.dt)
 
@@ -263,7 +263,7 @@ class Syphilis(ss.Infection):
 
                 # Birth outcomes must be modified to add probability of susceptible birth
                 birth_outcomes = self.pars.birth_outcomes[state]
-                assigned_outcomes = birth_outcomes.urvs(uids)-uids  # WHY??
+                assigned_outcomes = birth_outcomes.rvs(len(uids))-uids  # WHY?? # TODO: TEMP: NOT CRN SAFE
                 time_to_birth = -sim.people.age
 
                 # Schedule events