+++ starsim/states.py
@@ -9,7 +9,6 @@ import starsim as ss
 from starsim.settings import INT_NAN
 from starsim.settings import dtypes as sdt
 from numpy.lib.mixins import NDArrayOperatorsMixin  # Inherit from this to automatically gain operators like +, -, ==, <, etc.
-from scipy.stats._distn_infrastructure import rv_frozen
 
 
 __all__ = ['check_dtype', 'UIDArray', 'State', 'ArrayView']
@@ -450,7 +449,7 @@ class State(UIDArray):
 
     def _new_vals(self, uids):
         if isinstance(self.default, ss.Dist):
-            new_vals = self.default.urvs(uids)
+            new_vals = self.default.rvs(uids)
         elif callable(self.default):
             new_vals = self.default(len(uids))
         else: